# context/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Enter the access token returned by /api/auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "API Support",
      "url": "https://github.com/bruno-frascino/piggy-api"
    },
    "description": "A comprehensive stock portfolio management API built with Express and Prisma",
    "title": "Truffles API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tax-reports": {
      "get": {
        "parameters": [
          {
            "description": "Include earlier revisions that have been superseded by a regeneration",
            "in": "query",
            "name": "includeSuperseded",
            "required": false,
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report metadata (no PDF bytes/line items), newest financial year first"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List generated capital gains tax reports for the authenticated user",
        "tags": [
          "TaxReports"
        ]
      }
    },
    "/api/tax-reports/generate": {
      "post": {
        "description": "Computes a capital gains summary for the given Australian financial year across an explicit set of Trading Accounts (a \"declaration\"), renders a PDF, and stores it as a new revision for that (financial year, account selection) combination. Previous revisions are retained and stay downloadable so an already-lodged report is never overwritten. Regenerating with unchanged content returns the existing current revision instead of creating a new one.\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "accountIds": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "financialYearStartYear": {
                    "description": "e.g. 2025 for FY2025-26 (1 Jul 2025 - 30 Jun 2026)",
                    "type": "integer"
                  }
                },
                "required": [
                  "financialYearStartYear",
                  "accountIds"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Report generated"
          },
          "400": {
            "description": "Invalid input or no accounts selected"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Generate (or regenerate) an ATO capital gains tax report",
        "tags": [
          "TaxReports"
        ]
      }
    },
    "/api/tax-reports/position-usage": {
      "get": {
        "description": "Derived from each report's stored line items — no denormalised flag is kept. A usage entry is `stale` when the position or any of its transactions was modified after the report was generated, meaning the PDF no longer matches the underlying data and should be regenerated.\n",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "additionalProperties": {
                        "items": {
                          "properties": {
                            "financialYearLabel": {
                              "type": "string"
                            },
                            "generatedAt": {
                              "format": "date-time",
                              "type": "string"
                            },
                            "reportId": {
                              "type": "string"
                            },
                            "stale": {
                              "type": "boolean"
                            }
                          },
                          "type": "object"
                        },
                        "type": "array"
                      },
                      "type": "object"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Position id keyed map of report usages"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Map each position to the generated tax reports that already include it",
        "tags": [
          "TaxReports"
        ]
      }
    },
    "/api/tax-reports/{id}": {
      "delete": {
        "description": "Deleting the current revision promotes the most recent remaining revision back to current, so the carried-forward loss chain stays intact.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Report not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete a persisted tax report revision",
        "tags": [
          "TaxReports"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report detail including lineItems"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Report not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a single tax report's metadata and per-disposal line items",
        "tags": [
          "TaxReports"
        ]
      }
    },
    "/api/tax-reports/{id}/download": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "PDF file stream"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Report not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Download the persisted PDF for a tax report",
        "tags": [
          "TaxReports"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Development server",
      "url": "http://localhost:4000"
    }
  ],
...
}